MILAB-6497: extract package-builder build engine into package-builder-lib
