+++ lib/node/package-builder-lib/.gitignore
@@ -1,0 +1,3 @@
+pkg-*.tgz
+pkg-*.zip
+package.sw.json
