+++ lib/node/package-builder-lib/.oxlintrc.json
@@ -1,0 +1,3 @@
+{
+  "extends": ["node_modules/@milaboratories/ts-builder/dist/configs/oxlint-node.json"]
+}
