+++ lib/node/package-builder-lib/assets/conda-dockerfile.template
