+++ lib/node/package-builder-lib/assets/python-dockerfile.template
