+++ lib/node/package-builder-lib/src/__test__/package.json
