+++ lib/node/package-builder-lib/src/__test__/vite.config.mts
