+++ lib/node/package-builder-lib/src/core/__tests__/docker-entrypoint-generation.test.ts
