+++ lib/node/package-builder-lib/src/core/archive.ts
