+++ lib/node/package-builder-lib/src/core/conda/validate-spec.test.ts
