+++ lib/node/package-builder-lib/src/core/package-info.test.ts
