+++ lib/node/package-builder-lib/src/core/package-info.ts
