+++ lib/node/package-builder-lib/src/core/paths.ts
