+++ lib/node/package-builder-lib/src/core/resolver.ts
