+++ lib/node/package-builder-lib/src/core/schemas/artifacts.test.ts
