+++ lib/node/package-builder-lib/src/core/schemas/entrypoints.test.ts
