+++ lib/node/package-builder-lib/src/core/schemas/entrypoints.ts
