+++ lib/node/package-builder-lib/src/core/schemas/sw-json.ts
