+++ lib/node/package-builder-lib/src/core/schemas/test-artifacts.ts
