+++ lib/node/package-builder-lib/src/core/sw-json-render.ts
