+++ lib/node/package-builder-lib/src/defaults.ts
