+++ lib/node/package-builder-lib/src/requirements.txt
@@ -1,0 +1,1 @@
+# No dependencies specified
