+++ lib/node/package-builder-lib/tsconfig.json
@@ -1,0 +1,8 @@
+{
+  "extends": "@milaboratories/ts-configs/tsconfig.node.json",
+  "compilerOptions": {
+    "outDir": "./dist",
+    "rootDir": "./src"
+  },
+  "include": ["src"]
+}
