+++ lib/node/package-builder-lib/vitest.config.mts
@@ -1,0 +1,4 @@
+import { createVitestConfig } from "@milaboratories/build-configs";
+import { defineConfig } from "vitest/config";
+
+export default defineConfig(createVitestConfig());
