+++ tools/package-builder/src/commands/prepublish.ts
@@ -1,8 +1,6 @@
 import { Command } from "commander";
-import * as cmdOpts from "../core/cmd-opts";
-import * as util from "../core/util";
-import { Core } from "../core/core";
-import * as envs from "../core/envs";
+import * as cmdOpts from "../cmd-opts";
+import { util, envs, createBuilder } from "@platforma-sdk/package-builder-lib";
 
 export function prepublishCommand(): Command {
   const cmd = new Command("prepublish").description(
@@ -25,9 +23,9 @@ export function prepublishCommand(): Command {
     const flags = cmdOpts.toFlags(opts);
     const logger = util.createLogger(flags["log-level"]);
 
-    const core = new Core(logger, { packageRoot: flags["package-root"] });
+    const core = createBuilder(logger, { packageRoot: flags["package-root"] });
     core.buildMode = cmdOpts.modeFromFlag(flags.dev as cmdOpts.devModeName);
-    core.pkgInfo.version = flags.version;
+    core.version = flags.version;
     core.fullDirHash = flags["full-dir-hash"];
     core.allPlatforms = true;
