+++ tools/package-builder/src/commands/publish/all.ts
@@ -1,8 +1,6 @@
 import { Command } from "commander";
-import * as cmdOpts from "../../core/cmd-opts";
-import * as util from "../../core/util";
-import { Core } from "../../core/core";
-import * as envs from "../../core/envs";
+import * as cmdOpts from "../../cmd-opts";
+import { util, envs, createBuilder } from "@platforma-sdk/package-builder-lib";
 
 export function publishAllCommand(): Command {
   const cmd = new Command("all").description(
@@ -28,8 +26,8 @@ export function publishAllCommand(): Command {
     const flags = cmdOpts.toFlags(opts);
     const logger = util.createLogger(flags["log-level"]);
 
-    const core = new Core(logger, { packageRoot: flags["package-root"] });
-    core.pkgInfo.version = flags.version;
+    const core = createBuilder(logger, { packageRoot: flags["package-root"] });
+    core.version = flags.version;
     core.targetPlatform = flags.platform as util.PlatformType;
     core.allPlatforms = flags["all-platforms"];
