+++ tools/package-builder/src/commands/publish/docker.ts
@@ -1,8 +1,6 @@
 import { Command } from "commander";
-import * as cmdOpts from "../../core/cmd-opts";
-import * as util from "../../core/util";
-import { Core } from "../../core/core";
-import * as envs from "../../core/envs";
+import * as cmdOpts from "../../cmd-opts";
+import { util, envs, createBuilder } from "@platforma-sdk/package-builder-lib";
 
 export function publishDockerCommand(): Command {
   const cmd = new Command("docker").description("publish docker image to its registry");
@@ -24,8 +22,8 @@ export function publishDockerCommand(): Command {
     const flags = cmdOpts.toFlags(opts);
     const logger = util.createLogger(flags["log-level"]);
 
-    const core = new Core(logger, { packageRoot: flags["package-root"] });
-    core.pkgInfo.version = flags.version;
+    const core = createBuilder(logger, { packageRoot: flags["package-root"] });
+    core.version = flags.version;
 
     core.publishDockerImages({
       ids: flags["package-id"],
